+++ tests/microservices/Microservices.IsIdentifiable.Tests/ReviewerTests/SymbolsRulesFactoryTests.cs
@@ -9,13 +9,15 @@ namespace Microservices.IsIdentifiable.Tests.ReviewerTests
     public class SymbolsRulesFactoryTests
     {
         
-        [TestCase("MR Head 12-11-20","12-11-20",@"(\d\d-\d\d-\d\d)$")]
-        [TestCase("CT Head - 12/34/56","12/34/56",@"(\d\d/\d\d/\d\d)$")]
-        [TestCase("CT Head - 123-ABC-n4 fishfish","123-ABC-n4",@"(\d\d\d-[A-Z][A-Z][A-Z]-[a-z]\d)")]
-        [TestCase("123","123",@"^(\d\d\d)$")]
-        public void TestSymbols_OnePart(string input,string part, string expectedOutput)
+        [TestCase("MR Head 12-11-20","12-11-20",@"(\d\d-\d\d-\d\d)$",SymbolsRuleFactoryMode.Full)]
+        [TestCase("CT Head - 12/34/56","12/34/56",@"(\d\d/\d\d/\d\d)$",SymbolsRuleFactoryMode.Full)]
+        [TestCase("CT Head - 123-ABC-n4 fishfish","123-ABC-n4",@"(\d\d\d-[A-Z][A-Z][A-Z]-[a-z]\d)",SymbolsRuleFactoryMode.Full)]
+        [TestCase("CT Head - 123-ABC-n4 fishfish","123-ABC-n4",@"(123-[A-Z][A-Z][A-Z]-[a-z]4)",SymbolsRuleFactoryMode.CharactersOnly)]
+        [TestCase("CT Head - 123-ABC-n4 fishfish","123-ABC-n4",@"(\d\d\d-ABC-n\d)",SymbolsRuleFactoryMode.DigitsOnly)]
+        [TestCase("123","123",@"^(\d\d\d)$",SymbolsRuleFactoryMode.Full)]
+        public void TestSymbols_OnePart(string input,string part, string expectedOutput,SymbolsRuleFactoryMode mode)
         {
-            var f = new SymbolsRulesFactory();
+            var f = new SymbolsRulesFactory(){Mode = mode};
 
             var failure = new Failure(new[] {new FailurePart(part, FailureClassification.Person, input.IndexOf(part))})
             {
@@ -24,6 +26,7 @@ namespace Microservices.IsIdentifiable.Tests.ReviewerTests
 
             Assert.AreEqual(expectedOutput,f.GetPattern(this, failure));
         }
+
 
         [TestCase("12 Morton Street","12","eet",@"^(\d\d).*([a-z][a-z][a-z])$")]
         [TestCase("Morton MR Smith","MR","Smith",@"([A-Z][A-Z]).*([A-Z][a-z][a-z][a-z][a-z])$")]